Restrict numerical input, as accepted by the POST endpoint.
This will add a dependendy for Cargo. Update README.md to mention input restrictions. The endpoint will now return text instead of JSON, when out-of-range input was provided.

+++ src/lib/structs.rs
@@ -1,15 +1,18 @@
 use rocket::serde::Deserialize;
 use rocket::serde::Serialize;
+use validator::Validate;
 
-#[derive(Deserialize, Serialize)]
+#[derive(Deserialize, Serialize, Validate)]
 pub struct ApeIndexInput {
-    pub height: u32,
-    pub wingspan: u32,
+    #[validate(range(min = 1, max = 300, message = "Expected numerical input for height should be between 1 and 300"))]
+    pub height: i16,
+    #[validate(range(min = 1, max = 300, message = "Expected numerical input for wingspan should be between 1 and 300"))]
+    pub wingspan: i16,
 }
 
 #[derive(Deserialize, Serialize)]
 pub struct ApeIndexOutput {
-    pub height: u32,
-    pub wingspan: u32,
+    pub height: i16,
+    pub wingspan: i16,
     pub ape_index: f32,
 }
